Splitted variwide series into type files.
