+++ ts/Series/Variwide/VariwidePointOptions.d.ts
@@ -1,0 +1,34 @@
+/* *
+ *
+ *  Highcharts variwide module
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ * Imports
+ *
+ * */
+import type ColumnPointOptions from '../Column/ColumnPointOptions';
+
+/* *
+ *
+ * Declaration
+ *
+ * */
+export interface VariwidePointOptions extends ColumnPointOptions {
+    // Nothing here
+}
+
+/* *
+ *
+ * Export
+ *
+ * */
+export default VariwidePointOptions;
